refactor: split upstream_plugin and external into focused submodules
Break up two monolithic files for readability and testability:
- upstream_plugin/mod.rs -> types.rs, streaming.rs, mod.rs
- external.rs -> external/process.rs, streaming.rs, mod.rs

Adds 13 unit tests for previously untested logic including
PluginOutput deserialization, merge_body_into_arg, write_request
framing, and locate_* path resolution.
