+++ plenum-core/src/upstream_plugin/mod.rs
@@ -1,7 +1,8 @@
-use std::collections::HashMap;
+mod types;
 
-use bytes::Bytes;
-use plenum_js_runtime::StreamChunk;
+pub use types::*;
+
+use std::collections::HashMap;
 
 use crate::ctx::GatewayCtx;
 use crate::gateway_error::GatewayErrorResponse;
@@ -18,9 +19,7 @@ use plenum_config::ConfigValue;
 
 use crate::request_context::{ExtractionCtx, PingoraRequest};
 use pingora_proxy::Session;
-use serde::{Deserialize, Serialize};
 use tracing::Instrument;
-use ts_rs::TS;
 
 /// Log a plugin error and send a gateway error response, then return early.
 macro_rules! plugin_error_response {
@@ -37,71 +36,8 @@ macro_rules! plugin_error_response {
     }};
 }
 
-/// The request sub-object inside [`PluginInput`].
-/// Note: the request body is injected at the top level of the input by the JS
-/// runtime and is accessible as `input.body` in JavaScript.
-#[derive(Serialize, TS)]
-pub struct PluginRequest {
-    pub method: String,
-    /// The matched OpenAPI path template, e.g. `/users/{id}`.
-    pub route: String,
-    pub path: String,
-    /// Raw query string. Preserved for backward compatibility.
-    pub query: String,
-    /// Query parameters parsed according to the operation's OpenAPI parameter definitions.
-    /// Scalar values are type-coerced; arrays and objects follow the OAS style/explode rules.
-    /// Parameters not declared in the spec are included as raw strings.
-    #[serde(rename = "queryParams")]
-    #[ts(rename = "queryParams", type = "Record<string, unknown>")]
-    pub query_params: serde_json::Value,
-    pub headers: HashMap<String, String>,
-    #[ts(type = "Record<string, unknown>")]
-    pub params: HashMap<String, serde_json::Value>,
-}
-
-/// Input passed to a plugin's `handle()` function.
-#[derive(Serialize, TS)]
-pub struct PluginInput {
-    pub request: PluginRequest,
-    #[ts(type = "unknown")]
-    pub config: serde_json::Value,
-    #[ts(type = "unknown")]
-    pub operation: serde_json::Value,
-    #[ts(type = "Ctx")]
-    pub ctx: serde_json::Value,
-}
-
-/// Output returned by a plugin's `handle()` function.
-/// The response body is extracted separately by the JS runtime.
-#[derive(Deserialize, Default, TS)]
-pub struct PluginOutput {
-    #[ts(optional)]
-    pub status: Option<u16>,
-    /// Headers to set on the response. A `null` value removes the header.
-    #[ts(optional)]
-    pub headers: Option<HashMap<String, Option<String>>>,
-    #[ts(optional, type = "Record<string, unknown>")]
-    pub ctx: Option<serde_json::Map<String, serde_json::Value>>,
-}
-
-/// The actual shape of the input object that JS plugin `handle()` receives.
-/// Includes runtime-injected body fields not present on the base [`PluginInput`].
-///
-/// Used only for TypeScript type generation — never instantiated at runtime.
-#[derive(Serialize, TS)]
-#[ts(rename = "PluginInput")]
-pub struct JsPluginInput {
-    #[serde(flatten)]
-    #[ts(flatten)]
-    base: PluginInput,
-    /// The parsed request body, injected by the JS runtime.
-    #[ts(optional, type = "unknown")]
-    body: Option<serde_json::Value>,
-    /// Set to `"base64"` when `body` is a base64-encoded binary.
-    #[serde(rename = "bodyEncoding")]
-    #[ts(optional)]
-    body_encoding: Option<String>,
-}
+// Declared after the macro so `plugin_error_response!` is in scope.
+mod streaming;
 
 /// Convert `HashMap<String, Option<String>>` plugin headers into a flat
 /// `Vec<(String, String)>`, dropping entries with `None` values.
@@ -364,7 +300,7 @@ pub(crate) async fn dispatch(
         }
     }
 
-    // Step B: call plugin handle()
+    // Build the plugin input.
     let plugin_query_str = session.req_header().uri.query().unwrap_or("");
     let plugin_query_defs = oas_query::extract_query_params(&op.operation_meta);
     let plugin_query_params = serde_json::Value::Object(oas_query::parse_query_params(
@@ -415,249 +351,141 @@ pub(crate) async fn dispatch(
         &final_buf,
     );
 
-    // Step B: call plugin handle() — either streaming or full response
+    // Delegate to streaming or non-streaming path.
     if plugin.streaming {
-        // -----------------------------------------------------------------------
-        // Streaming path: call_stream returns chunks via mpsc channel.
-        // -----------------------------------------------------------------------
-        let (meta_value, mut rx) = match plugin
-            .runtime
-            .call_stream(
-                "handle",
-                serde_json::to_value(plugin_input).unwrap(),
-                js_body,
-                plugin.timeout,
-            )
-            .await
-        {
-            Ok(result) => result,
-            Err(e) => {
-                plugin_error_response!(session, ctx, "plugin handle() streaming error", e);
-            }
-        };
-
-        let parsed: PluginOutput = serde_json::from_value(meta_value.value).unwrap_or_default();
-        let mut plugin_status = parsed.status.unwrap_or(200);
-        let mut plugin_headers = parsed.headers.unwrap_or_default();
-        merge_ctx(&mut ctx.user_ctx, parsed.ctx);
-
-        // Step C: on_response interceptors (run normally — they don't use body)
-        run_on_response_interceptors(
-            op,
+        return streaming::dispatch_streaming(
+            session,
             ctx,
+            op,
+            plugin,
+            plugin_input,
+            js_body,
             &method,
             &route,
-            &mut plugin_status,
-            &mut plugin_headers,
         )
         .await;
+    }
 
-        // Step D: on_response_body interceptors are SKIPPED for streaming routes.
-        // Streaming has no full body buffer for interceptors to transform.
-        // Plugin author accepts this tradeoff when setting `streaming: true`.
-        if !op.interceptors.on_response_body.is_empty() {
-            log::warn!(
-                "on_response_body interceptors are not supported for streaming plugin route '{}' — skipped",
-                route,
-            );
-        }
-
-        // Step E: write header + stream chunks
-        let response_headers = flatten_plugin_headers(plugin_headers);
-        let mut resp_header =
-            pingora_http::ResponseHeader::build(plugin_status, None).map_err(|e| {
-                pingora_core::Error::because(
-                    pingora_core::ErrorType::InternalError,
-                    "build response header for streaming plugin",
-                    e,
-                )
-            })?;
-        for (name, value) in &response_headers {
-            resp_header
-                .insert_header(name.clone(), value.as_bytes())
-                .ok();
-        }
-        session
-            .write_response_header(Box::new(resp_header), false)
-            .await
-            .map_err(|e| {
-                pingora_core::Error::because(
-                    pingora_core::ErrorType::InternalError,
-                    "write response header for streaming plugin",
-                    e,
-                )
-            })?;
-
-        // Stream chunks until done or cancelled.
-        loop {
-            tokio::select! {
-                biased;
-                _ = ctx.cancellation.cancelled() => {
-                    // Client disconnected — stop reading chunks.
-                    break;
-                }
-                chunk = rx.recv() => {
-                    match chunk {
-                        Some(Ok(StreamChunk::Chunk(data))) => {
-                            session
-                                .write_response_body(Some(Bytes::from(data)), false)
-                                .await
-                                .map_err(|e| {
-                                    pingora_core::Error::because(
-                                        pingora_core::ErrorType::InternalError,
-                                        "write streaming plugin chunk",
-                                        e,
-                                    )
-                                })?;
-                        }
-                        Some(Ok(StreamChunk::Done)) | None => {
-                            // Finalize the chunked response.
-                            session
-                                .write_response_body(None, true)
-                                .await
-                                .map_err(|e| {
-                                    pingora_core::Error::because(
-                                        pingora_core::ErrorType::InternalError,
-                                        "finalize streaming plugin response",
-                                        e,
-                                    )
-                                })?;
-                            break;
-                        }
-                        Some(Err(e)) => {
-                            log::error!("plugin streaming error: {}", e);
-                            // Response header already sent — terminate chunked transfer.
-                            let _ = session.write_response_body(None, true).await;
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(true)
-    } else {
-        // -----------------------------------------------------------------------
-        // Non-streaming path: call returns full response body.
-        // -----------------------------------------------------------------------
-        let (mut plugin_status, mut plugin_headers, plugin_body, plugin_returned_ctx) = match plugin
-            .runtime
-            .call(
-                "handle",
-                serde_json::to_value(plugin_input).unwrap(),
-                js_body,
-                plugin.timeout,
+    // -----------------------------------------------------------------------
+    // Non-streaming path: call returns full response body.
+    // -----------------------------------------------------------------------
+    let (mut plugin_status, mut plugin_headers, plugin_body, plugin_returned_ctx) = match plugin
+        .runtime
+        .call(
+            "handle",
+            serde_json::to_value(plugin_input).unwrap(),
+            js_body,
+            plugin.timeout,
+        )
+        .await
+    {
+        Ok(output) => {
+            let parsed: PluginOutput = serde_json::from_value(output.value).unwrap_or_default();
+            (
+                parsed.status.unwrap_or(200),
+                parsed.headers.unwrap_or_default(),
+                output.body,
+                parsed.ctx,
             )
-            .await
-        {
-            Ok(output) => {
-                let parsed: PluginOutput = serde_json::from_value(output.value).unwrap_or_default();
-                (
-                    parsed.status.unwrap_or(200),
-                    parsed.headers.unwrap_or_default(),
-                    output.body,
-                    parsed.ctx,
-                )
-            }
-            Err(e) => {
-                plugin_error_response!(session, ctx, "plugin handle() error", e);
-            }
-        };
-        merge_ctx(&mut ctx.user_ctx, plugin_returned_ctx);
+        }
+        Err(e) => {
+            plugin_error_response!(session, ctx, "plugin handle() error", e);
+        }
+    };
+    merge_ctx(&mut ctx.user_ctx, plugin_returned_ctx);
 
-        // Step C: on_response interceptors
-        run_on_response_interceptors(
-            op,
-            ctx,
+    // Step C: on_response interceptors
+    run_on_response_interceptors(
+        op,
+        ctx,
+        &method,
+        &route,
+        &mut plugin_status,
+        &mut plugin_headers,
+    )
+    .await;
+
+    // Step D: on_response_body interceptors
+    let mut response_body_bytes = plugin_body.map(js_body_to_bytes).unwrap_or_default();
+    let response_content_type = plugin_headers
+        .get("content-type")
+        .and_then(|v| v.clone())
+        .clone();
+    for hook in &op.interceptors.on_response_body {
+        let js_body =
+            js_body_from_content_type(response_content_type.as_deref(), &response_body_bytes);
+        let effective_headers: HashMap<String, String> = plugin_headers
+            .iter()
+            .filter_map(|(k, v)| v.as_ref().map(|val| (k.clone(), val.clone())))
+            .collect();
+        let input = response_input_from_parts(
+            http::StatusCode::from_u16(plugin_status).unwrap_or(http::StatusCode::OK),
             &method,
             &route,
-            &mut plugin_status,
-            &mut plugin_headers,
+            &headers_hashmap_to_http_headermap(&effective_headers),
+            op.operation_meta.clone(),
+            ctx.rate_limit_state.clone(),
+            serde_json::Value::Object(ctx.user_ctx.clone()),
+        );
+        let mut input_json = serde_json::to_value(&input).unwrap();
+        merge_options(&mut input_json, hook.options.as_ref());
+        match call_interceptor(
+            hook.runtime.as_ref(),
+            &hook.function,
+            input_json,
+            js_body,
+            hook.timeout,
         )
-        .await;
-
-        // Step D: on_response_body interceptors
-        let mut response_body_bytes = plugin_body.map(js_body_to_bytes).unwrap_or_default();
-        let response_content_type = plugin_headers
-            .get("content-type")
-            .and_then(|v| v.clone())
-            .clone();
-        for hook in &op.interceptors.on_response_body {
-            let js_body =
-                js_body_from_content_type(response_content_type.as_deref(), &response_body_bytes);
-            let effective_headers: HashMap<String, String> = plugin_headers
-                .iter()
-                .filter_map(|(k, v)| v.as_ref().map(|val| (k.clone(), val.clone())))
-                .collect();
-            let input = response_input_from_parts(
-                http::StatusCode::from_u16(plugin_status).unwrap_or(http::StatusCode::OK),
-                &method,
-                &route,
-                &headers_hashmap_to_http_headermap(&effective_headers),
-                op.operation_meta.clone(),
-                ctx.rate_limit_state.clone(),
-                serde_json::Value::Object(ctx.user_ctx.clone()),
-            );
-            let mut input_json = serde_json::to_value(&input).unwrap();
-            merge_options(&mut input_json, hook.options.as_ref());
-            match call_interceptor(
-                hook.runtime.as_ref(),
-                &hook.function,
-                input_json,
-                js_body,
-                hook.timeout,
-            )
-            .await
-            {
-                Ok((
-                    InterceptorOutput::Continue {
-                        status,
-                        headers,
-                        ctx: returned_ctx,
-                    },
-                    body_out,
-                )) => {
-                    if let Some(s) = status {
-                        plugin_status = s;
-                    }
-                    if let Some(mods) = headers {
-                        for (k, v) in mods {
-                            match v {
-                                Some(val) => {
-                                    plugin_headers.insert(k, Some(val));
-                                }
-                                None => {
-                                    plugin_headers.remove(&k);
-                                }
+        .await
+        {
+            Ok((
+                InterceptorOutput::Continue {
+                    status,
+                    headers,
+                    ctx: returned_ctx,
+                },
+                body_out,
+            )) => {
+                if let Some(s) = status {
+                    plugin_status = s;
+                }
+                if let Some(mods) = headers {
+                    for (k, v) in mods {
+                        match v {
+                            Some(val) => {
+                                plugin_headers.insert(k, Some(val));
+                            }
+                            None => {
+                                plugin_headers.remove(&k);
                             }
                         }
                     }
-                    if let Some(b) = body_out {
-                        response_body_bytes = js_body_to_bytes(b);
-                    }
-                    merge_ctx(&mut ctx.user_ctx, returned_ctx);
                 }
-                Ok((InterceptorOutput::Respond { .. }, _)) => {
-                    log::warn!(
-                        "on_response_body interceptor returned 'respond' action for plugin route -- ignored"
-                    );
+                if let Some(b) = body_out {
+                    response_body_bytes = js_body_to_bytes(b);
                 }
-                Err(e) => {
-                    log::error!("on_response_body interceptor error: {}", e);
-                }
+                merge_ctx(&mut ctx.user_ctx, returned_ctx);
+            }
+            Ok((InterceptorOutput::Respond { .. }, _)) => {
+                log::warn!(
+                    "on_response_body interceptor returned 'respond' action for plugin route -- ignored"
+                );
+            }
+            Err(e) => {
+                log::error!("on_response_body interceptor error: {}", e);
             }
         }
-
-        // Step E: write response
-        let response_headers = flatten_plugin_headers(plugin_headers);
-        crate::proxy_utils::write_response(
-            session,
-            plugin_status,
-            &response_headers,
-            response_body_bytes,
-        )
-        .await?;
-
-        Ok(true)
     }
+
+    // Step E: write response
+    let response_headers = flatten_plugin_headers(plugin_headers);
+    crate::proxy_utils::write_response(
+        session,
+        plugin_status,
+        &response_headers,
+        response_body_bytes,
+    )
+    .await?;
+
+    Ok(true)
 }
